[flow] move dict_to_table() to dedicated utils module
